Anova Tables for Cumulative Link (Mixed) Models

These functions are methods for Anova to calculate type-II or type-III analysis-of-deviance tables for model objects produced by clm and clmm. Likelihood-ratio tests are calculated in both cases.

# S3 method for clm
Anova(mod, type = c("II", "III", 2, 3), ...)

# S3 method for clmm
Anova(mod, type = c("II", "III", 2, 3), ...)

Arguments

mod
clm or clmm object.

type
type of test, "II", "III", 2 or 3.

…
additional arguments to Anova. Not usable here.

Details

See help of the Anova for a detailed explanation of what "type II" and "typ III" mean.

Value

See Anova.

See also