Loadings of multivariate analyses

Returns loadings of a multivariate analysis.

MVA.load(x, xax = 1, yax = 2, set = c(12, 1, 2), space = 1, ...)

Arguments

x
a multivariate analysis (see Details).

xax
axis or axes for which to extract loadings.

yax
axis for which to extract loadings (ignored if length(xax) > 1).

set
variables to be displayed, when several sets are available (see Details). 12 (default) for both sets, 1 for X, 2 for Y.

space
variables to be displayed, when several spaces are available (see Details). space is the number of the space to be plotted.

...
not used.

Details

Many multivariate analyses are supported, from various packages:

- PCA: prcomp, princomp, dudi.pca, rda, pca, pca.

- sPCA: spca.

- IPCA: ipca.

- sIPCA: sipca.

- LDA: lda, discrimin.

- PLS-DA (PLS2 on a dummy-coded factor): plsda. X space only.

- sPLS-DA (sPLS2 on a dummy-coded factor): splsda. X space only.

- CPPLS: mvr. X space only.

- PLSR: mvr, pls, plsR (plsRglm package). X space only.

- sPLSR: pls. X space only.

- PLS-GLR: plsRglm (plsRglm package).

- PCR: mvr.

- CDA: discrimin, discrimin.coa.

- NSCOA: dudi.nsc.

- MCA: dudi.acm.

- Mix analysis: dudi.mix, dudi.hillsmith.

- PCIA: procuste. Set 1 is X, set 2 is Y.

- RDA (or PCAIV): pcaiv, pcaivortho, rda. With rda, space 1 is constrained space, space 2 is unconstrained space. Only constrained space is available with pcaiv, the opposite for pcaivortho.

- CCorA: rcc. Space 1 is X, space 2 is Y.

- rCCorA: rcc. Space 1 is X, space 2 is Y.

- CIA: coinertia. Space 1 is X, space 2 is Y.

- 2B-PLS: pls. Space 1 is X, space 2 is Y.

- 2B-sPLS: pls. Space 1 is X, space 2 is Y.

- rGCCA: wrapper.rgcca. Space can be 1 to n, the number of blocks (i.e. datasets).

- sGCCA: wrapper.sgcca. Space can be 1 to n, the number of blocks (i.e. datasets).

- DIABLO: block.plsda, block.splsda. Space can be 1 to n, the number of blocks (i.e. datasets).