Type II permutation MANOVA using distance matrices

This function is a wrapper to adonis but performs type II tests (whereas adonis performs type I).

adonis.II(formula, data, ...)

Arguments

formula
a typical model formula such as Y~A+B*C, but where Y is either a dissimilarity object (inheriting from class "dist") or data frame or a matrix; A, B, and C may be factors or continuous variables.

data
the data frame from which A, B and C would be drawn.

…
additional arguments to adonis. See help of this function.

Details

See adonis for detailed explanation of what is done. The only difference with adonis is that adonis.II performs type II tests instead of type I.

Value

a data frame of class "anova".

Examples

require(vegan)

#> Loading required package: vegan

#> Loading required package: permute

#> Loading required package: lattice

#> This is vegan 2.5-4

data(dune)
data(dune.env)

# Compare:
adonis(dune~Management*A1,data=dune.env,permutations=99)

#> 
#> Call:
#> adonis(formula = dune ~ Management * A1, data = dune.env, permutations = 99) 
#> 
#> Permutation: free
#> Number of permutations: 99
#> 
#> Terms added sequentially (first to last)
#> 
#>               Df SumsOfSqs MeanSqs F.Model      R2 Pr(>F)   
#> Management     3    1.4686 0.48953  3.2629 0.34161   0.01 **
#> A1             1    0.4409 0.44089  2.9387 0.10256   0.02 * 
#> Management:A1  3    0.5892 0.19639  1.3090 0.13705   0.17   
#> Residuals     12    1.8004 0.15003         0.41878          
#> Total         19    4.2990                 1.00000          
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1

adonis(dune~A1*Management,data=dune.env,permutations=99)

#> 
#> Call:
#> adonis(formula = dune ~ A1 * Management, data = dune.env, permutations = 99) 
#> 
#> Permutation: free
#> Number of permutations: 99
#> 
#> Terms added sequentially (first to last)
#> 
#>               Df SumsOfSqs MeanSqs F.Model      R2 Pr(>F)   
#> A1             1    0.7230 0.72295  4.8187 0.16817   0.01 **
#> Management     3    1.1865 0.39551  2.6362 0.27600   0.02 * 
#> A1:Management  3    0.5892 0.19639  1.3090 0.13705   0.19   
#> Residuals     12    1.8004 0.15003         0.41878          
#> Total         19    4.2990                 1.00000          
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1

# With:
adonis.II(dune~Management*A1,data=dune.env,permutations=99)

#> Permutation: free
#> Number of permutations: 99
#> 
#> Type II tests
#> Response: dune
#>               Sum Sq Mean Sq Df      F Pr(>F)   
#> Management    1.1865 0.39551  3 2.6362   0.01 **
#> A1            0.4409 0.44089  1 2.9387   0.03 * 
#> Management:A1 0.5892 0.19639  3 1.3090   0.18   
#> Residuals     1.8004 0.15003 12                 
#> Total         4.2990         19                 
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1

adonis.II(dune~A1*Management,data=dune.env,permutations=99)

#> Permutation: free
#> Number of permutations: 99
#> 
#> Type II tests
#> Response: dune
#>               Sum Sq Mean Sq Df      F Pr(>F)   
#> A1            0.4409 0.44089  1 2.9387   0.01 **
#> Management    1.1865 0.39551  3 2.6362   0.01 **
#> A1:Management 0.5892 0.19639  3 1.3090   0.32   
#> Residuals     1.8004 0.15003 12                 
#> Total         4.2990         19                 
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1