Significance test for the covariance between two datasets

Performs a permutation test based on the sum of square covariance between variables of two datasets, to test wether the (square) covariance is higher than expected under random association between the two datasets. The test is relevent parallel to a 2B-PLS.

cov.test(X, Y, scale.X = TRUE, scale.Y = TRUE, nperm = 999, progress = TRUE)

Arguments

X
a numeric vector, matrix or data frame.

Y
a numeric vector, matrix or data frame.

scale.X
logical, if TRUE (default) scaling of X is required.

scale.Y
logical, if TRUE (default) scaling of Y is required.

nperm
number of permutations.

progress
logical indicating if the progress bar should be displayed.

Details

The function deals with the limitted floating point precision, which can bias calculation of p-values based on a discrete test statistic distribution.

Value

method

a character string indicating the name of the test.

data.name

a character string giving the name(s) of the data, plus additional information.

statistic

the value of the test statistics.

permutations

the number of permutations.

p.value

the p-value of the test.