Deprecated functions in RVAideMemoire package

Functions that are not usable anymore, and will be entirely removed from the package in future versions.

back.lsmeans(...)
byf.normhist(...)
cor.sparse(...)
DA.confusion(...)
DA.valid(...)
DA.var(...)
dunn.test(...)
fc.multcomp(...)
friedman.rating.test(...)
kruskal.rating.test(...)
pairwise.manova(...)
pairwise.to.groups(...)
pairwise.wilcox.rating.test(...)
plot1comp.ind(...)
plot1comp.var(...)
PLSDA.ncomp(...)
PLSDA.test(...)
rating.lsmeans(...)
s.corcircle2(...)
scat.mix.categorical(...)
scat.mix.numeric(...)
scatter.coa2(...)
wilcox.paired.rating.multcomp(...)
wilcox.rating.signtest(...)
wilcox.rating.test(...)

Arguments

...
previous arguments.

Details

back.lsmeans and rating.lsmeans are replaced by back.emmeans and rating.emmeans. More generally, stop using package lsmeans and change to package emmeans, its new version.

byf.normhist was not very useful and byf.hist does nearly the same job.

cor.sparse is replaced by the more generic MVA.plot.

DA.confusion and DA.valid are replaced by the more generic MVA.cmv and MVA.cv.

DA.var is replaced by the more generic MVA.synt.

dunn.test is not very useful, prefer dunnTest from package FSA.

fc.multcomp is not useful anymore since emtrends (package emmeans) does the same job in a much more powerful manner (see argument var of emtrends).

friedman.rating.test, kruskal.rating.test, wilcox.rating.test, wilcox.rating.signtest, pairwise.wilcox.rating.test and wilcox.paired.rating.multcomp can be problematic with ratings (in which ties and zeroes are very frequent). The use of CLM(M)s (via clm and clmm) is recommended.

pairwise.manova is not useful anymore since emmeans (package emmeans) does the same job in a much more powerful manner (on "mlm" objects, created by lm and not manova)

pairwise.to.groups was not very useful.

plot1comp.ind, plot1comp.var, s.corcircle2, scat.mix.categorical, scat.mix.numeric and scatter.coa2 are replaced by the more generic MVA.plot.

PLSDA.ncomp was not really useful and mvr does nearly the same job.

PLSDA.test is replaced by the more generic MVA.test.