Least rectangles linear regression

Fits a least rectangle linear regression, possibly for each level of a factor.

least.rect(formula, data, conf.level = 0.95, theo = 1, adj = TRUE)

Arguments

formula
a formula of the form y ~ x, where y and x give the y and x variable, respectively. The formula can also be y ~ x | f to fit a (separate) regression for each level of the factor f.

data
an optional data frame containing the variables in the formula formula. By default the variables are taken from environment(formula).

conf.level
confidence level.

theo
theoretical value of the slope. If several regression are fitted, the same value is used for all comparisons of slope vs. theoretical value.

adj
logical indicating if, in case of several regressions fitted, confidence intervals and p-values should be Bonferroni-corrected for multiple testing.

Value

coefficients

regression parameters.

residuals

residuals.

fitted.values

fitted values.

call

the matched call.

model

the model frame used.

conf.level

confidence level.

conf.int

confidence interval of regression parameters.

theo

theoretical value of the slope.

comp

data frame of results for equality of the slope(s) to the theoretical value.

corr

data frame of results for significativity of the correlation coefficient(s).

multiple

logical, TRUE if several regressions are fitted.

adj

logical, TRUE if confidence intervals and p-values are corrected for multiple testing (only if several regressions are fitted).

Examples

x <- 1:30+rnorm(30,0,3)
y <- 1:30+rnorm(30,0,3)
regression1 <- least.rect(y~x)
summary(regression1)

#> 
#> Call:
#> least.rect(formula = y ~ x)
#> 
#> Residuals:
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>  -5.280  -1.477  -0.163   0.000   1.538   6.839 
#> 
#> Coefficients and 95% confidence interval:
#>                  inf   coeff    sup
#> (Intercept) -3.29970 -0.2844 2.2518
#> x            0.87771  1.0435 1.2406
#> 
#> Equality of the slope to 1:
#>   observed theoretical      Df      t Pr(>|t|)
#>     1.0435      1.0000 28.0000 0.5018   0.6197
#> 
#> Pearson's linear correlation coefficient (95% confidence interval):
#>        inf        r      sup       Df      t  Pr(>|t|)    
#>    0.78623  0.89343  0.94843 28.00000 10.524 3.077e-11 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 

x2 <- c(1:30,1:30)
y2 <- c(1:30+rnorm(30,0,3),seq(10,22,12/29)+rnorm(30,0,3))
fact <- gl(2,30,labels=LETTERS[1:2])
regression2 <- least.rect(y2~x2|fact)
summary(regression2)

#> 
#> Call:
#> least.rect(formula = y2 ~ x2 | fact)
#> 
#> Residuals:
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#> -7.4335 -1.7535  0.0403  0.0000  1.8012  6.0482 
#> 
#> Coefficients and Bonferroni-adjusted 95% confidence interval:
#>   (Intercept) 
#>         inf     coeff      sup
#> A -2.699111 0.3237466  2.87393
#> B  5.070669 8.2698086 10.54960
#>   x2 
#>         inf     coeff       sup
#> A 0.8876648 1.0521928 1.2472159
#> B 0.3647327 0.5118157 0.7182118
#> 
#> Equality of the slopes to 1 (Bonferroni-adjusted p-value):
#>   observed theoretical       Df      t  Pr(>|t|)    
#> A  1.05219     1.00000 28.00000 0.7056    0.9726    
#> B  0.51182     1.00000 28.00000 4.9452 6.443e-05 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> Pearson's linear correlation coefficients (Bonferroni-adjusted 95% confidence interval and p-value):
#>        inf        r      sup       Df       t  Pr(>|t|)    
#> A  0.82941  0.92428  0.96733 28.00000 12.8124 6.256e-13 ***
#> B  0.30992  0.63623  0.82845 28.00000  4.3638 0.0003148 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#>