EMMeans for Cumulative Link (Mixed) Models

Extracts EMMeans (produced by emmeans) from Cumulative Link (Mixed) Models (produced by clm or clmm), with different possible formats.

rating.emmeans(emm, type = c("prob", "cumprob", "class1", "class2"), level = 0.9)

Arguments

emm
object returned by emmeans applied on a clm or clmm object.

type
type of output to be returned: "prob" (default) gives probability of each rating, "cumprob" gives cumulative probabilities (Pi is probability to be <= to rating i), "class1" gives the most probable rating and "class2" gives the first rating for which the cumulative probability is >= to level.

level
used only for type "class2" (see type).

Details

A factor named cut must have been called in emmeans, to compute EMMeans per cut point (i.e. rating). Additionally, the argument mode of emmeans must have been set to "linear.predictor". Finally, the call to emmeans is typically like emmeans(model,~factor|cut,mode="linear.predictor") where factor is the factor (or interaction) giving levels for which EMMeans have to be computed.

See also

Examples

require(ordinal)

#> Loading required package: ordinal

#> 
#> Attaching package: 'ordinal'

#> The following object is masked _by_ '.GlobalEnv':
#> 
#>     wine

#> The following objects are masked from 'package:lme4':
#> 
#>     ranef, VarCorr

require(emmeans)

model <- clm(rating~contact*temp,data=wine)

#> Error in eval(predvars, data, env): object 'rating' not found

EMM <- emmeans(model,~contact:temp|cut,mode="linear.predictor")

#> Error in is(object, "emmGrid"): object 'model' not found

# Probabilities
rating.emmeans(EMM)

#> Error in rating.emmeans(EMM): object 'EMM' not found

# Cumulative probabilities
rating.emmeans(EMM,type="cumprob")

#> Error in rating.emmeans(EMM, type = "cumprob"): object 'EMM' not found

# Most probable rating
rating.emmeans(EMM,type="class1")

#> Error in rating.emmeans(EMM, type = "class1"): object 'EMM' not found