Divide into groups respecting relative proportions

Divides a data frame randomly, but respecting the relative proportions of levels of a factor in the original data frame. Each subset has roughly the same number of individuals, and the same relative proportions in respect to levels of the given factor.

splitf(set, fac, k)

Arguments

set
a data frame containing values to be divided into groups.

fac
a reference factor giving the relative proportions to be respected in each subset of set.

k
an integer giving the number of subsets to be generated.

Value

A list of subsets of set.

See also

Examples

data(iris)
iris2 <- iris[c(1:50,51:80,101:120),]

# Proportions to be respected
table(iris2$Species)/nrow(iris2)

#> 
#>     setosa versicolor  virginica 
#>        0.5        0.3        0.2 

# Splitting
result <- splitf(iris2,iris2$Species,3)

# All subsets have the same size
lapply(result,nrow)

#> [[1]]
#> [1] 34
#> 
#> [[2]]
#> [1] 33
#> 
#> [[3]]
#> [1] 33
#> 

# And respect the initial proportions
lapply(result,function(x) table(x$Species)/nrow(x))

#> [[1]]
#> 
#>     setosa versicolor  virginica 
#>  0.5000000  0.2941176  0.2058824 
#> 
#> [[2]]
#> 
#>     setosa versicolor  virginica 
#>  0.4848485  0.3030303  0.2121212 
#> 
#> [[3]]
#> 
#>     setosa versicolor  virginica 
#>  0.5151515  0.3030303  0.1818182 
#>