Wald tests for comparison of proportions to theoretical values

Performs pairwise comparisons of proportions to theoretical values.

wald.ptheo.multinom.test(x, p, p.method = "fdr")

Arguments

x
either a factor or a matrix with K columns giving the counts for each of the K classes.

p
theoretical proportions.

p.method
method for p-values correction. See help of p.adjust.

Details

The function builds K logistic regressions (in each case considering one class vs. the sum of all others) and uses wald.ptheo.test to test the hypothesis that the proportion of this class is equal to p[K].

Value

method

name of the test.

data.name

a character string giving the name(s) of the data.

observed

observed proportions.

expected

theoretical proportions.

p.adjust.method

method for p-values correction.

statistic

statistics of each test.

p.value2

corrected p-values.

p.value

data frame of results.

See also

Examples

response <- factor(rep(LETTERS[1:4],c(20,40,42,13)))
wald.ptheo.multinom.test(response,p=c(0.15,0.25,0.3,0.3))

#> 
#>         Pairwise comparisons using Wald tests
#> 
#> data:  x and p
#> 
#>    observed expected statistic   p-value    
#>  A   0.1739     0.15    0.7172 0.4732251    
#>  B   0.3478     0.25    2.3901 0.0369662   *
#>  C   0.3652     0.30    1.5140 0.1770439    
#>  D   0.1130     0.30   -4.1181 0.0001528 ***
#> 
#> P value adjustment method: fdr