Scores of multivariate analyses

Returns scores of a multivariate analysis.

MVA.scores(x, xax = 1, yax = 2, scaling = 2, set = c(12, 1, 2), space = 1, ...)

Arguments

x
a multivariate analysis (see Details).

xax
axis or axes for which to extract scores.

yax
axis for which to extract scores (ignored if length(xax) > 1).

scaling
type of scaling. Only available with some analyses performed with the vegan package. See Details.

set
scores to be displayed, when several sets are available (see Details). 12 (default) for both sets, 1 for rows or X, 2 for columns or Y.

space
scores to be displayed, when several spaces are available (see Details). space is the number of the space to be plotted.

…
not used.

Details

Many multivariate analyses are supported, from various packages:

- PCA: prcomp, princomp (if scores=TRUE), dudi.pca, rda, pca, pca. scaling can be defined for rda (see scores.rda).

- sPCA: spca.

- IPCA: ipca.

- sIPCA: sipca.

- PCoA: cmdscale (with at least on non-default argument), dudi.pco, wcmdscale (with at least one non-default argument), capscale, pco, pcoa.

- nMDS: monoMDS, metaMDS, nmds, isoMDS.

- LDA: lda, discrimin.

- PLS-DA (PLS2 on a dummy-coded factor): plsda. X space only.

- sPLS-DA (sPLS2 on a dummy-coded factor): splsda. X space only.

- CPPLS: mvr. X space only.

- PLSR: mvr, pls, plsR (plsRglm package). X space only.

- sPLSR: pls. X space only.

- PLS-GLR: plsRglm (plsRglm package).

- PCR: mvr.

- CDA: discrimin, discrimin.coa.

- NSCOA: dudi.nsc.

- MCA: dudi.acm.

- Mix analysis: dudi.mix, dudi.hillsmith.

- COA: dudi.coa, cca. Set 1 is rows, set 2 is columns. If set=12 (default), fac is not available and pch,cex, col can be defined differently for each set. scaling can be defined for cca (see scores.cca).

- DCOA: dudi.dec. Set 1 is rows, set 2 is columns. If set=12 (default), fac is not available and pch,cex, col can be defined differently for each set.

- PCIA: procuste. Set 1 is X, set 2 is Y. If set=12 (default), fac is not available and pch,cex, col can be defined differently for each set.

- Procrustean superimposition: procrustes. Set 1 is X, set 2 is Y. If set=12 (default), fac is not available and pch,cex, col can be defined differently for each set.

- GPA: GPA. Only the consensus ordination can be displayed.

- DPCoA: dpcoa. Set 1 is categories, set 2 is collections. If set=12 (default), fac is not available and pch,cex, col can be defined differently for each set.

- RDA (or PCAIV): pcaiv, pcaivortho, rda. With rda, space 1 is constrained space, space 2 is unconstrained space. Only constrained space is available with pcaiv, the opposite for pcaivortho. scaling can be defined for rda (see scores.rda).

- db-RDA (or CAP): capscale, dbrda. Space 1 is constrained space, space 2 is unconstrained space.

- CCA: pcaiv, cca. With rda, space 1 is constrained space, space 2 is unconstrained space. Only constrained space is available with pcaiv. Set 1 is rows, set 2 is columns. scaling can be defined for cca (see scores.cca).

- CCorA: CCorA, rcc. Space 1 is X, space 2 is Y. With rcc a third space is available, in which coordinates are means of X and Y coordinates.

- rCCorA: rcc. Space 1 is X, space 2 is Y, space 3 is a "common" space in which coordinates are means of X and Y coordinates.

- CIA: coinertia. Space 1 is X, space 2 is Y, space 3 is a "common" space where X and Y scores are normed. In space 3, set 1 is X and set 2 is Y. If set=12 in space 3 (default), fac is not available and pch,cex, col can be defined differently for each set.

- 2B-PLS: pls. Space 1 is X, space 2 is Y, space 3 is a "common" space in which coordinates are means of X and Y coordinates.

- 2B-sPLS: pls. Space 1 is X, space 2 is Y, space 3 is a "common" space in which coordinates are means of X and Y coordinates.

- rGCCA: rgcca, wrapper.rgcca. Space can be 1 to n, the number of blocks (i.e. datasets).

- sGCCA: sgcca, wrapper.sgcca. Space can be 1 to n, the number of blocks (i.e. datasets).

- DIABLO: block.plsda, block.splsda. Space can be 1 to n, the number of blocks (i.e. datasets).