Synthesis quality of multivariate analyses

Gives a simple estimator of the quality of the (descriptive) synthesis performed by a wide range of multivariate analyses.

MVA.synt(x, rows = 5)

Arguments

x
a multivariate analysis (see Details).

rows
maximum number of axes to print in the output.

Details

Many multivariate analyses are supported, from various packages.

- PCA: prcomp, princomp, dudi.pca, rda, pca, pca: % of total variance explained by each axis.

- sPCA: spca: % of total variance explained by each axis.

- IPCA: ipca: kurtosis of each axis.

- sIPCA: sipca: kurtosis of each axis.

- PCoA: cmdscale (with eig=TRUE), dudi.pco, wcmdscale (with eig=TRUE), capscale, pco, pcoa: % of total variance explained by each axis.

- nMDS: monoMDS, metaMDS, nmds, isoMDS: stress.

- RDA: pcaiv, pcaivortho, rda: % of constrained and unconstrained total variance, % of constrained variance explained by constrained axes (pcaiv and rda), % of unconstrained variance explained by unconstrained axes (pcaivortho and rda).

- db-RDA (or CAP): capscale, dbrda: % of constrained and unconstrained total variance, % of constrained variance explained by constrained axes, % of unconstrained variance explained by unconstrained axes.

- COA: dudi.coa, cca: % of total inertia explained by each axis.

- CCA: pcaiv, cca: % of constrained and unconstrained total inertia, % of constrained inertia explained by constrained axes, % of unconstrained inertia explained by unconstrained axes (cca only).

- CPPLS: mvr: % of X and Y variances explained by each axis.

- PLSR: mvr, plsR (plsRglm package): % of X and Y variances explained by each axis (only Y for the moment with plsR).

- 2B-PLS: pls: % of X/Y square covariance explained by each pair of axes, correlation between each pair of axes (canonical correlations).

- CCorA: CCorA, rcc: correlation between each pair of axes (canonical correlations).

- rCCorA: rcc: correlation between each pair of axes (canonical correlations).

- PCR: mvr: % of X and Y variances explained by each axis.

- MCA: dudi.acm: % of total inertia explained by each axis.

- Mix analysis: dudi.mix, dudi.hillsmith: % of total inertia explained by each axis.

- GPA: GPA: % of consensus and residual variance, % of total variance exlained by each axis, % of consensus variance explained by each axis, % of residual variance coming from each group of variables.

- RGCCA: rgcca, wrapper.rgcca: % of total intra-block variance explained by each axis, correlation between each pair of axes (canonical correlations).

- DIABLO: block.plsda, block.splsda: % of total intra-block variance explained by each axis, correlation between each pair of axes (canonical correlations).

- CIA: coinertia: RV coefficient, % of co-inertia explained by each pair of axes, correlation between each pair of axes (canonical correlations).

- PCIA: procuste: m2.

Examples

data(iris)
PCA <- prcomp(iris[,1:4])
MVA.synt(PCA)

#> Criterion: total variance (%)
#>  Axis Proportion Cumulative
#>     1      92.46      92.46
#>     2       5.31      97.77
#>     3       1.71      99.48
#>     4       0.52     100.00
#>